sample-langgraph/pyproject.toml
[project]
name = "sample-langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-core>=0.3.64",
    "langchain-openai>=0.3.21",
    "langgraph>=0.4.8",
    "python-dotenv>=1.1.0",
]
